fix: unblock the windows lib-test gate after the ownership re-bind
The ownership re-bind landed after the last local clippy run, so two
gate failures reached CI.

Drop a redundant isize cast in the new close test: LocalFixture::handle
already returns isize, so `as isize` trips -D clippy::unnecessary_cast
and fails the lib-test build.

Take the on-screen stage lock in the focused-filter test. It reads the
foreground twice - once through the filter, once to corroborate - and
this sub-phase is the first to add tests that deliberately raise their
own windows, so a sibling activation test landing between those two
reads made it fail for the sibling's reason. It stages nothing itself,
but the lock guards screen state rather than only screen real estate.
Verified by running the suite in isolation (passes) and in parallel
(failed before, three consecutive clean runs after).

+++ crates/windows/src/system/close_tests.rs
@@ -289,7 +289,7 @@ fn graceful_wm_close_skips_hwnd_whose_owner_pid_changed() {
     let token = process_identity::token_for_pid(pid)
         .expect("token")
         .expect("live");
-    super::post_wm_close_if_still_owned(foreign.handle() as isize, pid, &token)
+    super::post_wm_close_if_still_owned(foreign.handle(), pid, &token)
         .expect("foreign hwnd is skipped, not an error");
     assert!(
         process_still_alive(pid, &token),
